requirements.txt
awscli==1.11.17
boto3==1.4.4
botocore==1.5.55
click==6.6
colorama==0.3.7
Cython==0.25.2
docutils==0.12
enum34==1.1.6
fastparquet==0.0.6
Flask==0.11.1
funcsigs==1.0.2
futures==3.0.5
google-api-python-client==1.6.2
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
llvmlite==0.18.0
MarkupSafe==0.23
nose==1.3.7
numba==0.33.0
numpy==1.11.2
oauth2client==4.1.0
pandas==0.20.1
ply==3.10
py==1.4.33
pyasn1==0.2.3
pyasn1-modules==0.0.8
pytest==3.1.0
python-dateutil==2.5.3
pytz==2017.2
requests==2.10.0
rsa==3.4.2
s3transfer==0.1.10
singledispatch==3.4.0.3
six==1.10.0
TBB==0.1
thriftpy==0.3.9
uritemplate==3.0.0
virtualenv==15.0.2
Werkzeug==0.11.10
